Avoid reuse in azure pipelines job configuration
In the azure pipelines runs for #258 the jobs are failing for what looks
like a mismatch with the installed aer python package and what should
have been built. Just to rule that out for runs this commit makes 2
changes. First it creates a conda environment/virtualenv with a unique
name for each run. Secondly it runs git clean to remove any old built
wheels (assuming there are any). Hopefully this will provide more
reliability in future runs.