Add macos universal2 wheel builds to wheel jobs
This commit adds universal2 builds for macOS. The universal2 wheels will
enable running on arm64 macOS systems. As we don't have any arm64 mac
hardware available in CI systems yet these jobs are cross-compiling arm64
binaries from an x86_64 mac, we also can't test the arm64 binary
component as part of the job because of this.

This commit also takes the opportunity to update the cibuildwheel
version used for CI to the latest release 1.10.0, since we need to
use >=1.9.0 for arm on macOS support.